requirements.txt
lancedb>=0.12.0
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-typescript>=0.20.0
tree-sitter-java>=0.20.0
tree-sitter-cpp>=0.20.0
tree-sitter-c>=0.20.0
tree-sitter-c-sharp>=0.20.0
tree-sitter-rust>=0.20.0
tree-sitter-go>=0.20.0
openai>=1.0.0
click>=8.0.0
pydantic>=2.0.0
tqdm>=4.60.0
pathspec>=0.11.0
gitignore-parser>=0.1.0
PyYAML